dev-requirements.txt
pytest-django
django-extensions
pygraphviz
